+++ DarkStar.Engine/Ai/Humans/MushroomFinderAi.cs
@@ -15,6 +15,7 @@ using DarkStar.Api.World.Types.Npc;
 using DarkStar.Api.World.Types.Tiles;
 using DarkStar.Engine.Commands.Actions;
 using DarkStar.Network.Protocol.Messages.Common;
+using DarkStar.Network.Protocol.Messages.World;
 using Microsoft.Extensions.Logging;
 
 namespace DarkStar.Engine.Ai.Humans;
@@ -37,9 +38,11 @@ public class MushroomFinderAi : BaseAiBehaviourExecutor
             if (_currentStep >= _path.Count)
             {
                 Logger.LogInformation("MMMhh, yummy mushroom!");
+                await SendWorldMessageAsync("MMMhh, yummy mushroom!", WorldMessageType.Yell);
 
                 Engine.CommandService.EnqueueNpcAction(new GameObjectAction()
                 {
+                    MapId = MapId,
                     IsNpc = true,
                     NpcId = NpcGameObject.ID,
                     NpcObjectId = NpcEntity.Id,
@@ -74,7 +77,5 @@ public class MushroomFinderAi : BaseAiBehaviourExecutor
             _path = GetPathToPosition(mushrooms.PointPosition());
             _currentStep = 0;
         }
-
-
     }
 }