+++ DarkStar.Engine/Commands/Executors/GameObjectActionExecutor.cs
@@ -1,6 +1,7 @@
 using DarkStar.Api.Engine.Attributes.Commands;
 using DarkStar.Api.Engine.Commands.Base;
 using DarkStar.Api.Engine.Interfaces.Core;
+using DarkStar.Api.Engine.Map.Entities;
 using DarkStar.Api.Engine.Types.Commands;
 using DarkStar.Engine.Commands.Actions;
 using Microsoft.Extensions.Logging;
@@ -15,9 +16,20 @@ public class GameObjectActionExecutor : BaseCommandActionExecutor<GameObjectActi
 
     }
 
-    public override Task ProcessAsync(GameObjectAction action)
+    public override async Task ProcessAsync(GameObjectAction action)
     {
-
-        return Task.CompletedTask;
+        var worldObject = await Engine.WorldService.GetEntityByPositionAsync<WorldGameObject>(action.MapId, action.Position);
+        if (worldObject != null)
+        {
+            await Engine.ItemService.ExecuteGameObjectActionAsync(
+                worldObject,
+                action.MapId,
+                action.SessionId,
+                action.PlayerId,
+                action.IsNpc,
+                action.NpcId,
+                action.NpcObjectId
+            );
+        }
     }
 }