+++ DarkStar.Engine/Items/WorldObjects/MushroomGameObject.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using DarkStar.Api.Engine.Attributes.Objects;
+using DarkStar.Api.Engine.Interfaces.Core;
 using DarkStar.Api.Engine.Items.WorldObjects.Base;
+using DarkStar.Api.Engine.Map.Entities;
 using DarkStar.Api.World.Types.GameObjects;
 using Microsoft.Extensions.Logging;
 
@@ -14,8 +16,14 @@ namespace DarkStar.Engine.Items.WorldObjects;
 [GameObjectAction(GameObjectType.Prop_Mushroom)]
 public class MushroomGameObject : BaseWorldObjectAction
 {
-    public MushroomGameObject(ILogger<BaseWorldObjectAction> logger) : base(logger)
+    public MushroomGameObject(ILogger<BaseWorldObjectAction> logger, IDarkSunEngine engine) : base(logger, engine)
     {
 
+    }
+
+    public override ValueTask OnActivatedAsync(string mapId, WorldGameObject gameObject, Guid senderId, bool isNpc)
+    {
+        Logger.LogInformation("Mushroom activated");
+        return RemoveMySelfAsync();
     }
 }