+++ DarkStar.Engine/Items/WorldObjects/TestSchedulerObject.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using DarkStar.Api.Engine.Attributes.Objects;
+using DarkStar.Api.Engine.Interfaces.Core;
 using DarkStar.Api.Engine.Items.WorldObjects.Base;
 using DarkStar.Api.World.Types.GameObjects;
 using Microsoft.Extensions.Logging;
@@ -14,7 +15,7 @@ namespace DarkStar.Engine.Items.WorldObjects;
 [GameObjectAction(GameObjectType.Internal_World_Spawner)]
 public class TestSchedulerObject : BaseScheduledWorldObject
 {
-    public TestSchedulerObject(ILogger<BaseWorldObjectAction> logger) : base(logger)
+    public TestSchedulerObject(ILogger<BaseWorldObjectAction> logger, IDarkSunEngine engine) : base(logger, engine)
     {
         SetInterval(4000);
     }