+++ DarkStar.Engine/Services/JobSchedulerService.cs
@@ -1,12 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using DarkStar.Api.Attributes.Services;
 using DarkStar.Api.Engine.Interfaces.Services;
 using DarkStar.Engine.Services.Base;
-
 using FluentScheduler;
 using Humanizer;
 using Microsoft.Extensions.Logging;